Abstract front LED handlers from main kb code
